sql: move dist_sql to separate package
Moving dist_sql APIs to separate package. Adding `TestMain` to set
up the testing shim.

Renaming files and `DistSQL*` types to avoid stuttering.

Fixing `protobuf.mk` to only generate the extra stuff for the protos
in `roachpb`.
