+++ sql/distsql/api.proto
@@ -17,12 +17,12 @@
 // APIs for distributed SQL; see docs/RFCS/distributed_sql.md
 
 syntax = "proto2";
-package cockroach.sql;
-option go_package = "sql";
+package cockroach.distsql;
+option go_package = "distsql";
 
 import "cockroach/roachpb/data.proto";
 import "cockroach/roachpb/errors.proto";
-import "cockroach/sql/dist_sql_data.proto";
+import "cockroach/sql/distsql/data.proto";
 import weak "gogoproto/gogo.proto";
 
 message SetupFlowsRequest {
