+++ sql/distsql/tablereader_test.go
@@ -14,7 +14,7 @@
 //
 // Author: Radu Berinde (radu@cockroachlabs.com)
 
-package sql
+package distsql
 
 import (
 	"testing"
